+++ EnhanceQoL/Locales/enUS.lua
@@ -256,6 +256,8 @@ L["enableSquareMinimap"] = "Use a square minimap instead of the normal round"
 L["enableSquareMinimapDesc"] = "This option required a reload"
 L["showInstanceDifficulty"] = "Show instance difficulty"
 L["showInstanceDifficultyDesc"] = "Replace the default icon with a short text label (NM, HC, M, M+, LFR) showing the current instance difficulty."
+L["enableLandingPageMenu"] = "Enable Landing Page context menu"
+L["enableLandingPageMenuDesc"] = "Adds a right-click menu to expansion and garrison minimap buttons"
 -- L["instanceDifficultyUseIcon"] = "Use icons for difficulty"
 
 L["Profiles"] = "Profiles"